% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petalcluster-package.R
\docType{package}
\name{petalcluster-package}
\alias{petalcluster}
\alias{petalcluster-package}
\title{petalcluster: Density-Based Clustering via Rust}
\description{
R bindings to the Rust petal-clustering library providing DBSCAN, HDBSCAN, and OPTICS density-based clustering algorithms.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
