% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.petal}
\alias{plot.petal}
\alias{plot.petal_dbscan}
\alias{plot.petal_hdbscan}
\alias{plot.petal_optics}
\title{Plot clustering results}
\usage{
\method{plot}{petal_dbscan}(x, ...)

\method{plot}{petal_hdbscan}(x, ...)

\method{plot}{petal_optics}(x, ...)
}
\arguments{
\item{x}{A clustering result object.}

\item{...}{Additional arguments passed to \code{\link[=pairs]{pairs()}} or \code{\link[=plot.default]{plot.default()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Produces a scatter plot matrix (pairs plot) of clustered data, colored by
cluster assignment. For 2-column data a single scatter plot is produced
instead. Noise points (\code{NA} cluster) are shown as grey crosses.
}
