% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rings}
\alias{rings}
\title{rings}
\format{
A data frame with columns:
\describe{
\item{x}{X coordinate}
\item{y}{Y coordinate}
\item{cluster}{Cluster assignment}
}
}
\source{
Generated using a custom function.
}
\usage{
rings
}
\description{
A dataset of concentric rings with noise, used for testing clustering algorithms.
}
\keyword{datasets}
